Moved rest of sim files to sim/
