+++ sim/plots.py
